divide files in contracts
