+++ Source/TonSwift/Contracts/CommonMsgInfo/CommonMsgInfoExternalIn.swift
@@ -1,0 +1,12 @@
+//
+//  CommonMsgInfoExternalIn.swift
+//  
+//
+//  Created by Kirill Kirilenko on 20/07/2023.
+//
+
+public struct CommonMsgInfoExternalIn {
+    let src: ExternalAddress?
+    let dest: Address
+    let importFee: Coins
+}
