+++ Source/TonSwift/Contracts/Contract/ContractStateLast.swift
@@ -1,0 +1,14 @@
+//
+//  ContractStateLast.swift
+//  
+//
+//  Created by Kirill Kirilenko on 20/07/2023.
+//
+
+import Foundation
+import BigInt
+
+public struct ContractStateLast {
+    let lt: BigInt
+    let hash: Data
+}
