+++ Source/TonSwift/Contracts/Sender/Sender.swift
@@ -1,0 +1,4 @@
+public struct Sender {
+    let address: Address?
+    let send: ((SenderArguments) async throws -> Void)
+}
