+++ Source/TonSwift/Contracts/StateInit.swift
@@ -62,44 +62,3 @@ public struct StateInit: CellCodable {
         return StateInit(splitDepth: splitDepth, special: special, code: code, data: data, libraries: libraries)
     }
 }
-
-// Source: https://github.com/ton-blockchain/ton/blob/24dc184a2ea67f9c47042b4104bbb4d82289fac1/crypto/block/block.tlb#L139
-// tick_tock$_ tick:Bool tock:Bool = TickTock;
-
-struct TickTock: CellCodable {
-    var tick: Bool
-    var tock: Bool
-    
-    func storeTo(builder: Builder) throws {
-        try builder.store(bit: tick)
-            .store(bit: tock)
-    }
-    
-    static func loadFrom(slice: Slice) throws -> TickTock {
-        TickTock(
-            tick: try slice.loadBoolean(),
-            tock: try slice.loadBoolean()
-        )
-    }
-}
-
-// Source: https://github.com/ton-blockchain/ton/blob/24dc184a2ea67f9c47042b4104bbb4d82289fac1/crypto/block/block.tlb#L145
-// simple_lib$_ public:Bool root:^Cell = SimpleLib;
-
-public struct SimpleLibrary: Hashable {
-    var `public`: Bool
-    var root: Cell
-}
-
-extension SimpleLibrary: CellCodable {
-    public func storeTo(builder: Builder) throws {
-        try builder.store(bit: self.public)
-            .store(ref: root)
-    }
-    public static func loadFrom(slice: Slice) throws -> SimpleLibrary {
-        Self(
-            public: try slice.loadBoolean(),
-            root: try slice.loadRef()
-        )
-    }
-}
